Trait sc_rpc::author::AuthorApi

pub trait AuthorApi<Hash, BlockHash>: 'static + Send + Sync {
    type Metadata: PubSubMetadata;
    fn submit_extrinsic(
        &self, 
        extrinsic: Bytes
    ) -> Box<dyn Future<Item = Hash, Error = Error> + 'static + Send>;

    fn insert_key(
        &self, 
        key_type: String, 
        suri: String, 
        public: Bytes
    ) -> Result<(), Error>;

    fn rotate_keys(&self) -> Result<Bytes, Error>;

    fn pending_extrinsics(&self) -> Result<Vec<Bytes>, Error>;

    fn remove_extrinsic(
        &self, 
        bytes_or_hash: Vec<ExtrinsicOrHash<Hash>>
    ) -> Result<Vec<Hash>, Error>;

    fn watch_extrinsic(
        &self, 
        metadata: Self::Metadata, 
        subscriber: Subscriber<TransactionStatus<Hash, BlockHash>, Error>, 
        bytes: Bytes
    );

    fn unwatch_extrinsic(
        &self, 
        metadata: Option<Self::Metadata>, 
        id: SubscriptionId
    ) -> Result<bool, Error>;

    fn to_delegate(self) -> IoDelegate<Self, Self::Metadata>
    where
        BlockHash: Send + Sync + 'static + Serialize,
        Hash: Send + Sync + 'static + Serialize + DeserializeOwned,
    { ... }
}

Substrate authoring RPC API

Associated Types

type Metadata: PubSubMetadata

RPC metadata

Loading content...
Required methods

fn submit_extrinsic(
&self,
extrinsic: Bytes
) -> Box<dyn Future<Item = Hash, Error = Error> + 'static + Send>

Submit hex-encoded extrinsic for inclusion in block.

fn insert_key(
&self,
key_type: String,
suri: String,
public: Bytes
) -> Result<(), Error>

Insert a key into the keystore.

fn rotate_keys(&self) -> Result<Bytes, Error>

Generate new session keys and returns the corresponding public keys.

fn pending_extrinsics(&self) -> Result<Vec<Bytes>, Error>

Returns all pending extrinsics, potentially grouped by sender.

fn remove_extrinsic(
&self,
bytes_or_hash: Vec<ExtrinsicOrHash<Hash>>
) -> Result<Vec<Hash>, Error>

Remove given extrinsic from the pool and temporarily ban it to prevent reimporting.

fn watch_extrinsic(
&self,
metadata: Self::Metadata,
subscriber: Subscriber<TransactionStatus<Hash, BlockHash>, Error>,
bytes: Bytes
)

Submit an extrinsic to watch.

See TransactionStatus for details on transaction lifecycle.

fn unwatch_extrinsic(
&self,
metadata: Option<Self::Metadata>,
id: SubscriptionId
) -> Result<bool, Error>

Unsubscribe from extrinsic watching.

Loading content...
Provided methods

fn to_delegate(self) -> IoDelegate<Self, Self::Metadata> where
BlockHash: Send + Sync + 'static + Serialize,
Hash: Send + Sync + 'static + Serialize + DeserializeOwned,

Create an IoDelegate, wiring rpc calls to the trait methods.

Loading content...