Trait sc_rpc::author::AuthorApiServer

pub trait AuthorApiServer<Hash, BlockHash>: 'static + Send + Sync {
    fn submit_extrinsic<'life0, 'async_trait>(
        &'life0 self, 
        extrinsic: Bytes
    ) -> Pin<Box<dyn Future<Output = Result<Hash, Error>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn insert_key(
        &self, 
        key_type: String, 
        suri: String, 
        public: Bytes
    ) -> Result<(), Error>;
    fn rotate_keys(&self) -> Result<Bytes, Error>;
    fn has_session_keys(&self, session_keys: Bytes) -> Result<bool, Error>;
    fn has_key(
        &self, 
        public_key: Bytes, 
        key_type: String
    ) -> Result<bool, Error>;
    fn pending_extrinsics(&self) -> Result<Vec<Bytes, Global>, Error>;
    fn remove_extrinsic(
        &self, 
        bytes_or_hash: Vec<ExtrinsicOrHash<Hash>, Global>
    ) -> Result<Vec<Hash, Global>, Error>;
    fn watch_extrinsic(
        &self, 
        subscription_sink: PendingSubscription, 
        bytes: Bytes
    );

    fn into_rpc(self) -> RpcModule<Self>
    where
        Hash: 'static + Send + Sync + DeserializeOwned + Serialize,
        BlockHash: 'static + Send + Sync + Serialize,
    { ... }
}

Server trait implementation for the AuthorApi RPC API.

Required methods

fn submit_extrinsic<'life0, 'async_trait>(
&'life0 self,
extrinsic: Bytes
) -> Pin<Box<dyn Future<Output = Result<Hash, Error>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Submit hex-encoded extrinsic for inclusion in block.

fn insert_key(
&self,
key_type: String,
suri: String,
public: Bytes
) -> Result<(), Error>

Insert a key into the keystore.

fn rotate_keys(&self) -> Result<Bytes, Error>

Generate new session keys and returns the corresponding public keys.

fn has_session_keys(&self, session_keys: Bytes) -> Result<bool, Error>

Checks if the keystore has private keys for the given session public keys.

session_keys is the SCALE encoded session keys object from the runtime.

Returns true iff all private keys could be found.

fn has_key(&self, public_key: Bytes, key_type: String) -> Result<bool, Error>

Checks if the keystore has private keys for the given public key and key type.

Returns true if a private key could be found.

fn pending_extrinsics(&self) -> Result<Vec<Bytes, Global>, Error>

Returns all pending extrinsics, potentially grouped by sender.

fn remove_extrinsic(
&self,
bytes_or_hash: Vec<ExtrinsicOrHash<Hash>, Global>
) -> Result<Vec<Hash, Global>, Error>

Remove given extrinsic from the pool and temporarily ban it to prevent reimporting.

fn watch_extrinsic(&self, subscription_sink: PendingSubscription, bytes: Bytes)

Submit an extrinsic to watch.

See TransactionStatus for details on transaction life cycle.

Provided methods

fn into_rpc(self) -> RpcModule<Self> where
Hash: 'static + Send + Sync + DeserializeOwned + Serialize,
BlockHash: 'static + Send + Sync + Serialize,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.