Trait sc_rpc::chain::ChainApiClient

pub trait ChainApiClient<Number, Hash, Header, SignedBlock>: SubscriptionClientT where
    Number: 'static + Send + Sync,
    Hash: 'static + Send + Sync + Serialize + DeserializeOwned,
    Header: 'static + Send + Sync + DeserializeOwned,
    SignedBlock: 'static + Send + Sync + DeserializeOwned, {
    fn header<'life0, 'async_trait>(
        &'life0 self, 
        hash: Option<Hash>
    ) -> Pin<Box<dyn Future<Output = Result<Option<Header>, Error>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait + Sync,
    { ... }

    fn block<'life0, 'async_trait>(
        &'life0 self, 
        hash: Option<Hash>
    ) -> Pin<Box<dyn Future<Output = Result<Option<SignedBlock>, Error>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait + Sync,
    { ... }

    fn block_hash<'life0, 'async_trait>(
        &'life0 self, 
        hash: Option<ListOrValue<NumberOrHex>>
    ) -> Pin<Box<dyn Future<Output = Result<ListOrValue<Option<Hash>>, Error>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait + Sync,
    { ... }

    fn finalized_head<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Hash, Error>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait + Sync,
    { ... }

    fn subscribe_all_heads<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Subscription<Header>, Error>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait + Sync,
    { ... }

    fn subscribe_new_heads<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Subscription<Header>, Error>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait + Sync,
    { ... }

    fn subscribe_finalized_heads<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Subscription<Header>, Error>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait + Sync,
    { ... }
}

Client implementation for the ChainApi RPC API.

Provided methods

fn header<'life0, 'async_trait>(
&'life0 self,
hash: Option<Hash>
) -> Pin<Box<dyn Future<Output = Result<Option<Header>, Error>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait + Sync,

Get header.

fn block<'life0, 'async_trait>(
&'life0 self,
hash: Option<Hash>
) -> Pin<Box<dyn Future<Output = Result<Option<SignedBlock>, Error>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait + Sync,

Get header and body of a relay chain block.

fn block_hash<'life0, 'async_trait>(
&'life0 self,
hash: Option<ListOrValue<NumberOrHex>>
) -> Pin<Box<dyn Future<Output = Result<ListOrValue<Option<Hash>>, Error>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait + Sync,

Get hash of the n-th block in the canon chain.

By default returns latest block hash.

fn finalized_head<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Hash, Error>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait + Sync,

Get hash of the last finalized block in the canon chain.

fn subscribe_all_heads<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Subscription<Header>, Error>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait + Sync,

All head subscription.

fn subscribe_new_heads<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Subscription<Header>, Error>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait + Sync,

New head subscription.

fn subscribe_finalized_heads<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Subscription<Header>, Error>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait + Sync,

Finalized head subscription.