Trait sc_rpc::chain::ChainApiServer

pub trait ChainApiServer<Number, Hash, Header, SignedBlock>: Sized + Send + Sync + 'static {
    // Required methods
    fn header(&self, hash: Option<Hash>) -> Result<Option<Header>, Error>;
    fn block(&self, hash: Option<Hash>) -> Result<Option<SignedBlock>, Error>;
    fn block_hash(
        &self,
        hash: Option<ListOrValue<NumberOrHex>>
    ) -> Result<ListOrValue<Option<Hash>>, Error>;
    fn finalized_head(&self) -> Result<Hash, Error>;
    fn subscribe_all_heads(
        &self,
        subscription_sink: SubscriptionSink
    ) -> Result<(), SubscriptionEmptyError>;
    fn subscribe_new_heads(
        &self,
        subscription_sink: SubscriptionSink
    ) -> Result<(), SubscriptionEmptyError>;
    fn subscribe_finalized_heads(
        &self,
        subscription_sink: SubscriptionSink
    ) -> Result<(), SubscriptionEmptyError>;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self>
       where Number: Send + Sync + 'static,
             Hash: Send + Sync + 'static + DeserializeOwned + Serialize,
             Header: Send + Sync + 'static + Serialize,
             SignedBlock: Send + Sync + 'static + Serialize { ... }
}

Server trait implementation for the ChainApi RPC API.

Required Methods

fn header(&self, hash: Option<Hash>) -> Result<Option<Header>, Error>

Get header.

fn block(&self, hash: Option<Hash>) -> Result<Option<SignedBlock>, Error>

Get header and body of a block.

fn block_hash( &self, hash: Option<ListOrValue<NumberOrHex>> ) -> Result<ListOrValue<Option<Hash>>, Error>

Get hash of the n-th block in the canon chain.

By default returns latest block hash.

fn finalized_head(&self) -> Result<Hash, Error>

Get hash of the last finalized block in the canon chain.

fn subscribe_all_heads( &self, subscription_sink: SubscriptionSink ) -> Result<(), SubscriptionEmptyError>

All head subscription.

fn subscribe_new_heads( &self, subscription_sink: SubscriptionSink ) -> Result<(), SubscriptionEmptyError>

New head subscription.

fn subscribe_finalized_heads( &self, subscription_sink: SubscriptionSink ) -> Result<(), SubscriptionEmptyError>

Finalized head subscription.

Provided Methods

fn into_rpc(self) -> RpcModule<Self>where Number: Send + Sync + 'static, Hash: Send + Sync + 'static + DeserializeOwned + Serialize, Header: Send + Sync + 'static + Serialize, SignedBlock: Send + Sync + 'static + Serialize,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.