Trait sc_rpc::dev::DevApi

pub trait DevApi<Hash>: 'static + Send + Sync {
    fn block_stats(&self, block_hash: Hash) -> Result<Option<BlockStats>, Error>;

    fn to_delegate<M>(self) -> IoDelegate<Self, M>
    where
        M: Metadata,
        Hash: 'static + Send + Sync + DeserializeOwned,
    { ... }
}

Substrate dev API.

This API contains unstable and unsafe methods only meant for development nodes. They are all flagged as unsafe for this reason.

Required methods

fn block_stats(&self, block_hash: Hash) -> Result<Option<BlockStats>, Error>

Reexecute the specified block_hash and gather statistics while doing so.

This function requires the specified block and its parent to be available at the queried node. If either the specified block or the parent is pruned, this function will return None.

Provided methods

fn to_delegate<M>(self) -> IoDelegate<Self, M> where
M: Metadata,
Hash: 'static + Send + Sync + DeserializeOwned,

Create an IoDelegate, wiring rpc calls to the trait methods.