Trait sc_rpc::offchain::OffchainApiServer

pub trait OffchainApiServer: 'static + Sized + Send + Sync {
    fn set_local_storage(
        &self,
        kind: StorageKind,
        key: Bytes,
        value: Bytes
    ) -> Result<(), Error>;
    fn get_local_storage(
        &self,
        kind: StorageKind,
        key: Bytes
    ) -> Result<Option<Bytes>, Error>;

    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the OffchainApi RPC API.

Required Methods

fn set_local_storage(
&self,
kind: StorageKind,
key: Bytes,
value: Bytes
) -> Result<(), Error>

Set offchain local storage under given key and prefix.

fn get_local_storage(
&self,
kind: StorageKind,
key: Bytes
) -> Result<Option<Bytes>, Error>

Get offchain local storage under given key and prefix.

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.