Trait sc_rpc::state::ChildStateApi

pub trait ChildStateApi<Hash>: 'static + Send + Sync {
    type Metadata: Metadata;
    pub fn storage_keys(
        &self, 
        child_storage_key: PrefixedStorageKey, 
        prefix: StorageKey, 
        hash: Option<Hash>
    ) -> Box<dyn Future<Item = Vec<StorageKey, Global>, Error = Error> + 'static + Send, Global>;

    pub fn storage(
        &self, 
        child_storage_key: PrefixedStorageKey, 
        key: StorageKey, 
        hash: Option<Hash>
    ) -> Box<dyn Future<Item = Option<StorageData>, Error = Error> + 'static + Send, Global>;

    pub fn storage_hash(
        &self, 
        child_storage_key: PrefixedStorageKey, 
        key: StorageKey, 
        hash: Option<Hash>
    ) -> Box<dyn Future<Item = Option<Hash>, Error = Error> + 'static + Send, Global>;

    pub fn storage_size(
        &self, 
        child_storage_key: PrefixedStorageKey, 
        key: StorageKey, 
        hash: Option<Hash>
    ) -> Box<dyn Future<Item = Option<u64>, Error = Error> + 'static + Send, Global>;

    pub fn to_delegate(self) -> IoDelegate<Self, Self::Metadata>
    where
        Hash: Send + Sync + 'static + Serialize + DeserializeOwned,
    { ... }
}

Substrate child state API

Note that all PrefixedStorageKey are desierialized from json and not guaranted valid.

Associated Types

type Metadata: Metadata

RPC Metadata

Loading content...
Required methods

pub fn storage_keys(
&self,
child_storage_key: PrefixedStorageKey,
prefix: StorageKey,
hash: Option<Hash>
) -> Box<dyn Future<Item = Vec<StorageKey, Global>, Error = Error> + 'static + Send, Global>

Returns the keys with prefix from a child storage, leave empty to get all the keys

pub fn storage(
&self,
child_storage_key: PrefixedStorageKey,
key: StorageKey,
hash: Option<Hash>
) -> Box<dyn Future<Item = Option<StorageData>, Error = Error> + 'static + Send, Global>

Returns a child storage entry at a specific block’s state.

pub fn storage_hash(
&self,
child_storage_key: PrefixedStorageKey,
key: StorageKey,
hash: Option<Hash>
) -> Box<dyn Future<Item = Option<Hash>, Error = Error> + 'static + Send, Global>

Returns the hash of a child storage entry at a block’s state.

pub fn storage_size(
&self,
child_storage_key: PrefixedStorageKey,
key: StorageKey,
hash: Option<Hash>
) -> Box<dyn Future<Item = Option<u64>, Error = Error> + 'static + Send, Global>

Returns the size of a child storage entry at a block’s state.

Loading content...
Provided methods

pub fn to_delegate(self) -> IoDelegate<Self, Self::Metadata> where
Hash: Send + Sync + 'static + Serialize + DeserializeOwned,

Create an IoDelegate, wiring rpc calls to the trait methods.

Loading content...