Trait sc_rpc::state::StateApi

pub trait StateApi<Hash>: 'static + Send + Sync {
    type Metadata: PubSubMetadata;
    fn call(
        &self, 
        name: String, 
        bytes: Bytes, 
        hash: Option<Hash>
    ) -> Box<dyn Future<Error = Error, Item = Bytes> + 'static + Send>;

    fn storage_keys(
        &self, 
        prefix: StorageKey, 
        hash: Option<Hash>
    ) -> Box<dyn Future<Error = Error, Item = Vec<StorageKey>> + 'static + Send>;

    fn storage(
        &self, 
        key: StorageKey, 
        hash: Option<Hash>
    ) -> Box<dyn Future<Error = Error, Item = Option<StorageData>> + 'static + Send>;

    fn storage_hash(
        &self, 
        key: StorageKey, 
        hash: Option<Hash>
    ) -> Box<dyn Future<Error = Error, Item = Option<Hash>> + 'static + Send>;

    fn storage_size(
        &self, 
        key: StorageKey, 
        hash: Option<Hash>
    ) -> Box<dyn Future<Error = Error, Item = Option<u64>> + 'static + Send>;

    fn child_storage_keys(
        &self, 
        child_storage_key: StorageKey, 
        child_info: StorageKey, 
        child_type: u32, 
        prefix: StorageKey, 
        hash: Option<Hash>
    ) -> Box<dyn Future<Error = Error, Item = Vec<StorageKey>> + 'static + Send>;

    fn child_storage(
        &self, 
        child_storage_key: StorageKey, 
        child_info: StorageKey, 
        child_type: u32, 
        key: StorageKey, 
        hash: Option<Hash>
    ) -> Box<dyn Future<Error = Error, Item = Option<StorageData>> + 'static + Send>;

    fn child_storage_hash(
        &self, 
        child_storage_key: StorageKey, 
        child_info: StorageKey, 
        child_type: u32, 
        key: StorageKey, 
        hash: Option<Hash>
    ) -> Box<dyn Future<Error = Error, Item = Option<Hash>> + 'static + Send>;

    fn child_storage_size(
        &self, 
        child_storage_key: StorageKey, 
        child_info: StorageKey, 
        child_type: u32, 
        key: StorageKey, 
        hash: Option<Hash>
    ) -> Box<dyn Future<Error = Error, Item = Option<u64>> + 'static + Send>;

    fn metadata(
        &self, 
        hash: Option<Hash>
    ) -> Box<dyn Future<Error = Error, Item = Bytes> + 'static + Send>;

    fn runtime_version(
        &self, 
        hash: Option<Hash>
    ) -> Box<dyn Future<Error = Error, Item = RuntimeVersion> + 'static + Send>;

    fn query_storage(
        &self, 
        keys: Vec<StorageKey>, 
        block: Hash, 
        hash: Option<Hash>
    ) -> Box<dyn Future<Error = Error, Item = Vec<StorageChangeSet<Hash>>> + 'static + Send>;

    fn subscribe_runtime_version(
        &self, 
        metadata: Self::Metadata, 
        subscriber: Subscriber<RuntimeVersion, Error>
    );

    fn unsubscribe_runtime_version(
        &self, 
        metadata: Option<Self::Metadata>, 
        id: SubscriptionId
    ) -> Result<bool, Error>;

    fn subscribe_storage(
        &self, 
        metadata: Self::Metadata, 
        subscriber: Subscriber<StorageChangeSet<Hash>, Error>, 
        keys: Option<Vec<StorageKey>>
    );

    fn unsubscribe_storage(
        &self, 
        metadata: Option<Self::Metadata>, 
        id: SubscriptionId
    ) -> Result<bool, Error>;

    fn to_delegate(self) -> IoDelegate<Self, Self::Metadata>
    where
        Hash: Send + Sync + 'static + Serialize + DeserializeOwned,
    { ... }
}

Substrate state API

Associated Types

type Metadata: PubSubMetadata

RPC Metadata

Loading content...
Required methods

fn call(
&self,
name: String,
bytes: Bytes,
hash: Option<Hash>
) -> Box<dyn Future<Error = Error, Item = Bytes> + 'static + Send>

Call a contract at a block's state.

fn storage_keys(
&self,
prefix: StorageKey,
hash: Option<Hash>
) -> Box<dyn Future<Error = Error, Item = Vec<StorageKey>> + 'static + Send>

Returns the keys with prefix, leave empty to get all the keys

fn storage(
&self,
key: StorageKey,
hash: Option<Hash>
) -> Box<dyn Future<Error = Error, Item = Option<StorageData>> + 'static + Send>

Returns a storage entry at a specific block's state.

fn storage_hash(
&self,
key: StorageKey,
hash: Option<Hash>
) -> Box<dyn Future<Error = Error, Item = Option<Hash>> + 'static + Send>

Returns the hash of a storage entry at a block's state.

fn storage_size(
&self,
key: StorageKey,
hash: Option<Hash>
) -> Box<dyn Future<Error = Error, Item = Option<u64>> + 'static + Send>

Returns the size of a storage entry at a block's state.

fn child_storage_keys(
&self,
child_storage_key: StorageKey,
child_info: StorageKey,
child_type: u32,
prefix: StorageKey,
hash: Option<Hash>
) -> Box<dyn Future<Error = Error, Item = Vec<StorageKey>> + 'static + Send>

Returns the keys with prefix from a child storage, leave empty to get all the keys

fn child_storage(
&self,
child_storage_key: StorageKey,
child_info: StorageKey,
child_type: u32,
key: StorageKey,
hash: Option<Hash>
) -> Box<dyn Future<Error = Error, Item = Option<StorageData>> + 'static + Send>

Returns a child storage entry at a specific block's state.

fn child_storage_hash(
&self,
child_storage_key: StorageKey,
child_info: StorageKey,
child_type: u32,
key: StorageKey,
hash: Option<Hash>
) -> Box<dyn Future<Error = Error, Item = Option<Hash>> + 'static + Send>

Returns the hash of a child storage entry at a block's state.

fn child_storage_size(
&self,
child_storage_key: StorageKey,
child_info: StorageKey,
child_type: u32,
key: StorageKey,
hash: Option<Hash>
) -> Box<dyn Future<Error = Error, Item = Option<u64>> + 'static + Send>

Returns the size of a child storage entry at a block's state.

fn metadata(
&self,
hash: Option<Hash>
) -> Box<dyn Future<Error = Error, Item = Bytes> + 'static + Send>

Returns the runtime metadata as an opaque blob.

fn runtime_version(
&self,
hash: Option<Hash>
) -> Box<dyn Future<Error = Error, Item = RuntimeVersion> + 'static + Send>

Get the runtime version.

fn query_storage(
&self,
keys: Vec<StorageKey>,
block: Hash,
hash: Option<Hash>
) -> Box<dyn Future<Error = Error, Item = Vec<StorageChangeSet<Hash>>> + 'static + Send>

Query historical storage entries (by key) starting from a block given as the second parameter.

NOTE This first returned result contains the initial state of storage for all keys. Subsequent values in the vector represent changes to the previous state (diffs).

fn subscribe_runtime_version(
&self,
metadata: Self::Metadata,
subscriber: Subscriber<RuntimeVersion, Error>
)

New runtime version subscription

fn unsubscribe_runtime_version(
&self,
metadata: Option<Self::Metadata>,
id: SubscriptionId
) -> Result<bool, Error>

Unsubscribe from runtime version subscription

fn subscribe_storage(
&self,
metadata: Self::Metadata,
subscriber: Subscriber<StorageChangeSet<Hash>, Error>,
keys: Option<Vec<StorageKey>>
)

New storage subscription

fn unsubscribe_storage(
&self,
metadata: Option<Self::Metadata>,
id: SubscriptionId
) -> Result<bool, Error>

Unsubscribe from storage subscription

Loading content...
Provided methods

fn to_delegate(self) -> IoDelegate<Self, Self::Metadata> where
Hash: Send + Sync + 'static + Serialize + DeserializeOwned,

Create an IoDelegate, wiring rpc calls to the trait methods.

Loading content...