sc_rpc::system

Struct System

pub struct System<B: Block> { /* private fields */ }

System API implementation

Implementations

impl<B: Block> System<B>

pub fn new( info: SystemInfo, send_back: TracingUnboundedSender<Request<B>>, ) -> Self

Creates new System.

The send_back will be used to transmit some of the requests. The user is responsible for reading from that channel and answering the requests.