Trait sc_rpc::system::SystemApi

pub trait SystemApi<Hash, Number>: 'static + Send + Sync {
    fn system_name(&self) -> Result<String, Error>;

    fn system_version(&self) -> Result<String, Error>;

    fn system_chain(&self) -> Result<String, Error>;

    fn system_type(&self) -> Result<ChainType, Error>;

    fn system_properties(&self) -> Result<Map<String, Value>, Error>;

    fn system_health(&self) -> Receiver<Health>;

    fn system_local_peer_id(&self) -> Receiver<String>;

    fn system_local_listen_addresses(&self) -> Receiver<Vec<String>>;

    fn system_peers(
        &self
    ) -> Compat<Pin<Box<dyn Future<Output = Result<Vec<PeerInfo<Hash, Number>>, Error>> + 'static + Send>>>;

    fn system_network_state(
        &self
    ) -> Compat<Pin<Box<dyn Future<Output = Result<Value, Error>> + 'static + Send>>>;

    fn system_add_reserved_peer(
        &self, 
        peer: String
    ) -> Compat<Pin<Box<dyn Future<Output = Result<(), Error>> + 'static + Send>>>;

    fn system_remove_reserved_peer(
        &self, 
        peer_id: String
    ) -> Compat<Pin<Box<dyn Future<Output = Result<(), Error>> + 'static + Send>>>;

    fn system_node_roles(&self) -> Receiver<Vec<NodeRole>>;

    fn to_delegate<M>(self) -> IoDelegate<Self, M>
    where
        Hash: Send + Sync + 'static + Serialize,
        M: Metadata,
        Number: Send + Sync + 'static + Serialize,
    { ... }
}

Substrate system RPC API

Required methods

fn system_name(&self) -> Result<String, Error>

Get the node's implementation name. Plain old string.

fn system_version(&self) -> Result<String, Error>

Get the node implementation's version. Should be a semver string.

fn system_chain(&self) -> Result<String, Error>

Get the chain's name. Given as a string identifier.

fn system_type(&self) -> Result<ChainType, Error>

Get the chain's type.

fn system_properties(&self) -> Result<Map<String, Value>, Error>

Get a custom set of properties as a JSON object, defined in the chain spec.

fn system_health(&self) -> Receiver<Health>

Return health status of the node.

Node is considered healthy if it is:

connected to some peers (unless running in dev mode)

not performing a major sync

fn system_local_peer_id(&self) -> Receiver<String>

Returns the base58-encoded PeerId of the node.

fn system_local_listen_addresses(&self) -> Receiver<Vec<String>>

Returns the multiaddresses that the local node is listening on

The addresses include a trailing /p2p/ with the local PeerId, and are thus suitable to be passed to system_addReservedPeer or as a bootnode address for example.

fn system_peers(
&self
) -> Compat<Pin<Box<dyn Future<Output = Result<Vec<PeerInfo<Hash, Number>>, Error>> + 'static + Send>>>

Returns currently connected peers

fn system_network_state(
&self
) -> Compat<Pin<Box<dyn Future<Output = Result<Value, Error>> + 'static + Send>>>

Returns current state of the network.

Warning: This API is not stable.

fn system_add_reserved_peer(
&self,
peer: String
) -> Compat<Pin<Box<dyn Future<Output = Result<(), Error>> + 'static + Send>>>

Adds a reserved peer. Returns the empty string or an error. The string parameter should encode a p2p multiaddr.

/ip4/198.51.100.19/tcp/30333/p2p/QmSk5HQbn6LhUwDiNMseVUjuRYhEtYj4aUZ6WfWoGURpdV is an example of a valid, passing multiaddr with PeerId attached.

fn system_remove_reserved_peer(
&self,
peer_id: String
) -> Compat<Pin<Box<dyn Future<Output = Result<(), Error>> + 'static + Send>>>

Remove a reserved peer. Returns the empty string or an error. The string should encode only the PeerId e.g. QmSk5HQbn6LhUwDiNMseVUjuRYhEtYj4aUZ6WfWoGURpdV.

fn system_node_roles(&self) -> Receiver<Vec<NodeRole>>

Returns the roles the node is running as.

Loading content...
Provided methods

fn to_delegate<M>(self) -> IoDelegate<Self, M> where
Hash: Send + Sync + 'static + Serialize,
M: Metadata,
Number: Send + Sync + 'static + Serialize,

Create an IoDelegate, wiring rpc calls to the trait methods.

Loading content...