Crate sc_rpc

Substrate RPC implementation.

A core implementation of Substrate RPC interfaces.

Modules

author

Substrate block-author/full-node API.

chain

Substrate blockchain API.

dev

Implementation of the DevApiServer trait providing debug utilities for Substrate based blockchains.

offchain

Substrate offchain API.

state

Substrate state API.

system

Substrate system API.

Structs

RandomIntegerSubscriptionId

Generates random integers as subscription ID.

RandomStringSubscriptionId

Generates random strings of length len as subscription ID.

Enums

DenyUnsafe

Signifies whether a potentially unsafe RPC should be denied.

Traits

RpcSubscriptionIdProvider

Trait to generate subscription IDs.

Type Definitions

SubscriptionTaskExecutor

Task executor that is being used by RPC subscriptions.