Trait sc_rpc::chain::ChainApiServer

pub trait ChainApiServer<Number, Hash, Header, SignedBlock>: 'static + Send + Sync {
    fn header<'life0, 'async_trait>(
        &'life0 self, 
        hash: Option<Hash>
    ) -> Pin<Box<dyn Future<Output = Result<Option<Header>, Error>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn block<'life0, 'async_trait>(
        &'life0 self, 
        hash: Option<Hash>
    ) -> Pin<Box<dyn Future<Output = Result<Option<SignedBlock>, Error>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn block_hash(
        &self, 
        hash: Option<ListOrValue<NumberOrHex>>
    ) -> Result<ListOrValue<Option<Hash>>, Error>;

    fn finalized_head(&self) -> Result<Hash, Error>;

    fn subscribe_all_heads(&self, subscription_sink: PendingSubscription);

    fn subscribe_new_heads(&self, subscription_sink: PendingSubscription);

    fn subscribe_finalized_heads(&self, subscription_sink: PendingSubscription);

    fn into_rpc(self) -> RpcModule<Self>
    where
        Number: 'static + Send + Sync,
        Hash: 'static + Send + Sync + DeserializeOwned + Serialize,
        Header: 'static + Send + Sync + Serialize,
        SignedBlock: 'static + Send + Sync + Serialize,
    { ... }
}

Server trait implementation for the ChainApi RPC API.

Required methods

fn header<'life0, 'async_trait>(
&'life0 self,
hash: Option<Hash>
) -> Pin<Box<dyn Future<Output = Result<Option<Header>, Error>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Get header.

fn block<'life0, 'async_trait>(
&'life0 self,
hash: Option<Hash>
) -> Pin<Box<dyn Future<Output = Result<Option<SignedBlock>, Error>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Get header and body of a relay chain block.

fn block_hash(
&self,
hash: Option<ListOrValue<NumberOrHex>>
) -> Result<ListOrValue<Option<Hash>>, Error>

Get hash of the n-th block in the canon chain.

By default returns latest block hash.

fn finalized_head(&self) -> Result<Hash, Error>

Get hash of the last finalized block in the canon chain.

fn subscribe_all_heads(&self, subscription_sink: PendingSubscription)

All head subscription.

fn subscribe_new_heads(&self, subscription_sink: PendingSubscription)

New head subscription.

fn subscribe_finalized_heads(&self, subscription_sink: PendingSubscription)

Finalized head subscription.

Provided methods

fn into_rpc(self) -> RpcModule<Self> where
Number: 'static + Send + Sync,
Hash: 'static + Send + Sync + DeserializeOwned + Serialize,
Header: 'static + Send + Sync + Serialize,
SignedBlock: 'static + Send + Sync + Serialize,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.